+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixed a JavaScript error that occurred when creating a new Dashboard widget. ([#19479](https://github.com/craftcms/cms/issues/19479))
 - Fixed a bug where users without a local password could not start an elevated session using OAuth. ([#19512](https://github.com/craftcms/cms/pull/19512))
 - Fixed a bug where Yii log targets configured via `config/craft/app.php` did not receive messages logged with `Craft::info()` and related methods. ([#19517](https://github.com/craftcms/cms/pull/19517))
+- Moved filesystem URL settings to `CraftCms\Cms\Filesystem\Filesystems\Local`, with legacy filesystem compatibility remaining available through `craftcms/yii2-adapter`. ([#19525](https://github.com/craftcms/cms/pull/19525))
 
 ## 6.0.0-alpha.17 - 2026-08-18